+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 5.1.0
+## Changes
+- Introduced Rails autoinstrumentation for Controller Actions.  Only
+  applied if Rails is present and the configuration parameter
+  `instrument_rails_controller_actions` is set to true.
+
 # Version 5.0.0
 ## Breaking changes
 - Renamed `Hallmonitor::Outputters::DogstatsdOutputter` to